requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.5
django-heroku==0.3.1
django-rest-framework==0.1.0
django-summernote==0.8.11.6
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jsonapi==2.4.0
djoser==2.0.3
gunicorn==20.0.4
idna==2.7
inflection==0.4.0
psycopg2==2.8.5
PyJWT==1.6.4
python-dotenv==0.10.3
pytz==2020.1
requests==2.19.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
whitenoise==5.1.0
